Set default text_encoder_quantization to bf16 in wgp.py

+++ wgp.py
@@ -2114,7 +2114,7 @@ if not Path(config_load_filename).is_file():
         "attention_mode" : "auto",  
         "transformer_types": [], 
         "transformer_quantization": "int8",
-        "text_encoder_quantization" : "int8",
+        "text_encoder_quantization" : "bf16",
         "save_path": "outputs",  
         "image_save_path": "outputs",  
         "compile" : "true",
